# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ortus Source Query API
  description: |
    REST API für die Abfrage von Datenquellen mittels Punktkoordinaten.

    Ortus ermöglicht die effiziente räumliche Abfrage von Datenquellen und unterstützt
    verschiedene Koordinatensysteme (SRIDs). Die API kann mehrere Datenquellen gleichzeitig
    verwalten und abfragen.

    ## Koordinatensysteme

    Die API unterstützt verschiedene Koordinatensysteme über den `srid` Parameter:
    - **4326** (WGS84): Standard GPS-Koordinaten (lon/lat)
    - **3857** (Web Mercator): Web-Mapping Standard
    - **25832/25833** (ETRS89/UTM): Deutscher/EU Standard
    - **31466/31467** (DHDN/Gauß-Krüger): Deutsches Legacy-System

    ## Authentifizierung

    Derzeit ist keine Authentifizierung erforderlich.
  version: 1.0.0
  contact:
    name: Ortus API Support
    url: https://github.com/jobrunner/ortus
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: API Version 1
  - url: /
    description: Root (für Health-Endpoints)
...
paths:
...
  /gazetteer:
    get:
      tags:
        - Gazetteer
      summary: Reverse-Geocoding + Peilung
      description: |
        Löst eine Koordinate auf ihre Verwaltungshierarchie (`admin`) auf und berechnet
        eine Peilung zum salientesten nahen Ort (`bearing`, z. B. "4 km E Würzburg").
        Jeder Teil ist `null`, wenn es kein Ergebnis gibt — keine Admin-Abdeckung bzw.
        kein Anker in Reichweite. Nur verfügbar, wenn das Gazetteer-Feature aktiv ist.
        Ein projizierter `srid` (z. B. 3857) wird intern nach WGS84 reprojiziert; die
        Antwort enthält die reprojizierte Koordinate im `wgs84`-Block. Nur ein `srid`,
        der sich nicht nach WGS84 transformieren lässt, wird mit 422 abgelehnt.
      operationId: gazetteer
      parameters:
        - $ref: '#/components/parameters/LonParam'
        - $ref: '#/components/parameters/LatParam'
        - $ref: '#/components/parameters/XParam'
        - $ref: '#/components/parameters/YParam'
        - $ref: '#/components/parameters/SridParam'
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GazetteerResponse'
              example:
                coordinate:
                  x: 9.93
                  y: 49.79
                  srid: 4326
                wgs84:
                  lon: 9.93
                  lat: 49.79
                admin:
                  country_iso: DE
                  hierarchy:
                    - level: 8
                      name: Würzburg
                      name_native: ""          # empty: already-Latin name (identical to name)
                      name_source: latin-osm
                      equivalent: municipality
                      local_term: Kreisfreie Stadt
                      equivalent_description: Local municipal authority (city/town/commune).
                    - level: 4
                      name: Bayern
                      name_native: ""          # empty: already-Latin name (identical to name)
                      name_source: latin-osm
                      equivalent: state
                      local_term: Land
                      equivalent_description: First-level subdivision (state, region, province).
                islands: null             # null: Würzburg is inland (an array of islands when the point is on one)
                bearing:
                  reference: Würzburg
                  name_native: ""            # empty: already-Latin name (identical to reference)
                  name_source: latin-osm
                  class: city
                  distance_km: 4.0
                  azimuth: 90.0
                  compass: E
                  label: "4 km E Würzburg"
                  inside: false
                exposure:                 # null when no DEM is wired or the point/neighbour lacks coverage
                  slope_deg: 12.5
                  slope_percent: 22.2
                  aspect_deg: 135.0       # null when flat
                  aspect_compass: SE      # "" when flat
                  flat: false
                  sample_spacing_m: 30.0
                  source:
                    name: "Copernicus DEM GLO-30"
                    url: "https://doi.org/10.5270/ESA-c5d3d65"
                    attribution: "© DLR e.V. 2010-2014 und © Airbus Defence and Space GmbH 2014-2018 provided under COPERNICUS by the European Union and ESA; all rights reserved."
                elevation:                # null when no DEM is configured
                  meters: 177.0
                  accuracy_m: 4.0
                  accuracy_basis: "GLO-30 LE90 (absolute)"
                  horizontal_accuracy_m: 6.0
                  vertical_datum: "EGM2008"
                  sea_level: false
                  surface_model: "DSM"
                  source:
                    name: "Copernicus DEM GLO-30"
                    url: "https://doi.org/10.5270/ESA-c5d3d65"
                    attribution: "© DLR e.V. 2010-2014 und © Airbus Defence and Space GmbH 2014-2018 provided under COPERNICUS by the European Union and ESA; all rights reserved."
                sources:
                  - code: latin-osm
                    short: OSM name (already Latin)
                    long: Taken verbatim from the OpenStreetMap name tag; already Latin script, no transliteration applied.
                    standard: ""    # latin-osm has no romanization standard (null in the manifest)
                license:
                  name: "ODbL-1.0"
                  url: "https://opendatacommons.org/licenses/odbl/1-0/"
                  attribution: "© OpenStreetMap contributors (ODbL 1.0); Natural Earth (public domain); GeoNames (CC BY 4.0); NGA GNS (public domain)"
        '400':
          description: Ungültige Parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
...
    LonParam:
      name: lon
      in: query
      description: Longitude (Längengrad) für WGS84-Koordinaten
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
      example: 13.405
    LatParam:
      name: lat
      in: query
      description: Latitude (Breitengrad) für WGS84-Koordinaten
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
      example: 52.52
    XParam:
      name: x
      in: query
      description: X-Koordinate für andere Koordinatensysteme
      schema:
        type: number
        format: double
      example: 389283
    YParam:
      name: y
      in: query
      description: Y-Koordinate für andere Koordinatensysteme
      schema:
        type: number
        format: double
      example: 5819450
    SridParam:
      name: srid
      in: query
      description: |
        Spatial Reference ID der Eingabekoordinaten.
        Standard ist 4326 (WGS84).
      schema:
        type: integer
        default: 4326
      example: 4326
...
  schemas:
    GazetteerResponse:
      type: object
      description: Reverse-Geocoding- und Peilungs-Ergebnis für eine Koordinate.
      allOf:
        - type: object
          required:
            - coordinate
          properties:
            coordinate:
              $ref: '#/components/schemas/Coordinate'
            wgs84:
              allOf:
                - $ref: '#/components/schemas/Wgs84Coordinate'
              nullable: true
              description: >-
                Der Abfragepunkt in WGS84 (lon/lat); bei projiziertem `srid`
                reprojiziert (siehe with-gazetteer / Endpoint-Beschreibung).
        - $ref: '#/components/schemas/GazetteerData'
    GazetteerData:
      type: object
      description: >-
        Gazetteer-Anreicherung für eine Koordinate: Verwaltungshierarchie (admin),
        Insel(n) (islands), Peilung (bearing), Geländeexposition (exposure), Höhe
        über NN (elevation), name_source-Erläuterungen (sources) und
        Datensatz-Lizenz/Attribution (license). Geliefert von GET /api/v1/gazetteer
        und — standardmäßig — im `gazetteer`-Block von GET /api/v1/query.
      properties:
        admin:
          nullable: true
          type: object
          description: Verwaltungshierarchie, die die Koordinate enthält (null ohne Abdeckung).
          properties:
            country_iso:
              type: string
              description: ISO 3166-1 alpha-2 Ländercode.
            hierarchy:
              type: array
              description: Ebenen, lokalste zuerst.
              items:
                type: object
                properties:
                  level:
                    type: integer
                    description: OSM admin_level.
                  name:
                    type: string
                    description: Name in lateinischer Schrift (immer gesetzt).
                  name_native:
                    type: string
                    description: Name in Originalschrift (leer, wenn identisch mit name).
                  name_source:
                    type: string
                    description: >-
                      Codierung der Namensherkunft/Romanisierung; im
                      response-weiten sources-Block beschrieben. Leer, wenn keine.
                  equivalent:
                    type: string
                    description: Semantische Bedeutung (country | state | … | municipality).
                  local_term:
                    type: string
                    description: Landesspezifischer Begriff für diese Ebene (z. B. „Landkreis").
                  equivalent_description:
                    type: string
                    description: Generische Beschreibung des equivalent-Levels.
        islands:
          nullable: true
          type: array
          description: >-
            Insel(n), deren Polygon die Koordinate enthält (eigener Layer,
            unabhängig von der Verwaltungsabdeckung aufgelöst). null, wenn der
            Punkt auf keiner erfassten Insel liegt oder kein Insel-Layer
            konfiguriert ist. Mehrere Einträge bei verschachtelten Inseln.
          items:
            type: object
            properties:
              name:
                type: string
                description: Inselname in lateinischer Schrift (immer gesetzt).
              name_native:
                type: string
                description: Inselname in Originalschrift (leer, wenn identisch mit name).
              name_source:
                type: string
                description: >-
                  Codierung der Namensherkunft/Romanisierung; im response-weiten
                  sources-Block beschrieben. Leer, wenn keine.
        bearing:
          nullable: true
          type: object
          description: Peilung zum salientesten nahen Ort (null ohne Anker in Reichweite).
          properties:
            reference:
              type: string
              description: Name des Anker-Orts (lateinische Schrift).
            name_native:
              type: string
              description: Name des Anker-Orts in Originalschrift (leer, wenn identisch).
            name_source:
              type: string
              description: >-
                Codierung der Namensherkunft/Romanisierung des Ankers; im
                response-weiten sources-Block beschrieben. Leer, wenn keine.
            class:
              type: string
              description: OSM place-Klasse (village | town | city).
            distance_km:
              type: number
              format: double
            azimuth:
              type: number
              format: double
              description: Azimut Referenz→Punkt in Grad (0=N, 90=O).
            compass:
              type: string
              description: Quantisierte Himmelsrichtung (z. B. E, NW). Leer, wenn `inside`.
            label:
              type: string
              description: >-
                Fertiges Label: „4 km E Würzburg", „prope Würzburg" (nah, außerhalb)
                oder „in Würzburg" (`inside`). Präfixe folgen der Etiketten-Konvention:
                lateinisch „in" bzw. „prope" (= „bei/nahe", etablierter Fundort-Terminus).
            inside:
              type: boolean
              description: >-
                Der Abfragepunkt liegt INNERHALB der Verwaltungseinheit des Anker-Orts
                — wir sind IN diesem Ort („in X"), nicht nur in der Nähe („prope X").
                Per Containment bestimmt (nicht per Distanz), gilt also auch weit vom
                Ortsmittelpunkt großer Städte. Ein Client kann bei `true` die Peilung
                weglassen (der Fund liegt direkt im Ort).
        exposure:
          nullable: true
          type: object
          description: >-
            Geländeexposition am Abfragepunkt: Hangneigung (slope) und
            Expositionsrichtung (aspect), aus dem DEM per Horn-Verfahren über eine
            3×3-Nachbarschaft. null, wenn kein DEM konfiguriert ist oder der Punkt
            (bzw. ein Nachbar) keine DEM-Abdeckung hat. Da die Quelle ein DSM ist
            (Oberfläche inkl. Vegetation/Bebauung), spiegeln Neigung/Richtung über
            Wald/Bebauung die Kronen/Dächer, nicht das Gelände; auf flachem Grund
            ist die Richtung rauschbestimmt (siehe `flat`).
          properties:
            slope_deg:
              type: number
              format: double
              description: Hangneigung in Grad (0 = eben, 90 = senkrecht).
            slope_percent:
              type: number
              format: double
              description: Hangneigung als Prozent-Gefälle (100·tan(slope)).
            aspect_deg:
              type: number
              format: double
              nullable: true
              description: >-
                Expositions-Azimut in Grad (0=N, 90=O, im Uhrzeigersinn), also die
                Richtung, in die der Hang abfällt. null, wenn `flat` (Richtung
                unbestimmt).
            aspect_compass:
              type: string
              description: >-
                Quantisierte Exposition (N, NE, …, 8-Punkt-Rose). Leer, wenn `flat`.
            flat:
              type: boolean
              description: >-
                true, wenn die Neigung unter der Flach-Schwelle (~2°) liegt; die
                Expositionsrichtung ist dann rauschbestimmt und wird weggelassen.
            sample_spacing_m:
              type: number
              format: double
              description: >-
                Horizontaler Abstand der Gradient-Stützpunkte in Metern (die
                Horn-Basislänge ist das Doppelte); richtet sich nach der
                DEM-Auflösung (GLO-30 ≈ 30 m).
            source:
              type: object
              nullable: true
              description: DEM-Quelle (Lizenz/Attribution), wie bei elevation.
              properties:
                name: { type: string }
                url: { type: string }
                attribution: { type: string }
        elevation:
          nullable: true
          type: object
          description: >-
            Höhe über NN am Abfragepunkt, aus einem kontinuierlichen Höhen-Raster
            (DEM). null, wenn kein DEM konfiguriert ist. `sea_level: true` bedeutet,
            dass keine DEM-Kachel den Punkt abdeckt (Ozean/außerhalb der Abdeckung);
            `meters` ist dann 0. Für abgeleitete Größen (Hangneigung/Exposition) ist
            der *relative* Fehler maßgeblich — siehe accuracy_basis.
          properties:
            meters:
              type: number
              format: double
              description: Orthometrische Höhe über NN in Metern.
            accuracy_m:
              type: number
              format: double
              description: >-
                Vertikale Genauigkeit in Metern. Aktuell eine datensatzweite
                Konstante (LE90); künftig per-Punkt aus der Height Error Mask (HEM).
            accuracy_basis:
              type: string
              description: >-
                Grundlage von accuracy_m, z. B. "GLO-30 LE90 (absolute)" oder
                "Copernicus HEM (per-pixel 1σ)".
            horizontal_accuracy_m:
              type: number
              format: double
              description: Horizontale Genauigkeit in Metern (LE90).
            vertical_datum:
              type: string
              description: Vertikaler Bezug, z. B. "EGM2008".
            sea_level:
              type: boolean
              description: >-
                true, wenn keine DEM-Kachel den Punkt abdeckt; meters ist dann 0
                (Meeresspiegel-Konvention).
            surface_model:
              type: string
              description: >-
                Oberflächenmodell-Hinweis, z. B. "DSM" (Oberfläche inkl.
                Vegetation/Bebauung, nicht bare-earth).
            source:
              type: object
              nullable: true
              description: >-
                Lizenz/Attribution der DEM-Quelle (z. B. Copernicus GLO-30),
                getrennt von der Gazetteer-license. Muss ein Client anzeigen.
              properties:
                name:
                  type: string
                  description: Lizenzname (SPDX-Id oder Bezeichnung).
                url:
                  type: string
                  description: Link zum Lizenztext bzw. zur Quelle.
                attribution:
                  type: string
                  description: Anzuzeigender Attributionstext.
        sources:
          type: array
          description: >-
            Response-weiter Herkunfts-Auszug: jede in admin/bearing vorkommende
            name_source-Codierung wird hier genau einmal beschrieben (nicht pro
            Datensatz wiederholt). Leeres Array, wenn keine Codierungen gesetzt sind.
          items:
            type: object
            properties:
              code:
                type: string
                description: Codierung, auf die name_source verweist.
              short:
                type: string
                description: Kurzbezeichnung der Quelle/Codierung.
              long:
                type: string
                description: Ausführliche Beschreibung der Quelle/Codierung.
              standard:
                type: string
                description: Zugrundeliegender Standard bzw. Zitierhinweis.
        license:
          type: object
          nullable: true
          description: >-
            Datensatzweite Lizenz/Attribution der Gazetteer-Daten (z. B. OSM/ODbL,
            GeoNames, Natural Earth), aus dem Manifest (ortus-gazetteer.yaml). Fehlt,
            wenn im Manifest keine Lizenz gesetzt ist. Diese Attribution muss ein
            Client anzeigen.
          properties:
            name:
              type: string
              description: Lizenzname (SPDX-Id oder Bezeichnung), z. B. "ODbL-1.0".
            url:
              type: string
              description: Link zum Lizenztext.
            attribution:
              type: string
              description: Anzuzeigender Attributionstext.
    Wgs84Coordinate:
      type: object
      description: >-
        Eine WGS84-Koordinate (EPSG:4326) als lon/lat. Explizit geografisch (nicht
        x/y/srid), damit nachgelagerte Services direkt damit rechnen/speichern können.
      properties:
        lon:
          type: number
          format: double
          description: Längengrad (WGS84).
        lat:
          type: number
          format: double
          description: Breitengrad (WGS84).
      required:
        - lon
        - lat
    Coordinate:
      type: object
      description: Geografische Koordinate
      properties:
        x:
          type: number
          format: double
          description: X-Koordinate (Longitude bei WGS84)
        y:
          type: number
          format: double
          description: Y-Koordinate (Latitude bei WGS84)
        srid:
          type: integer
          description: Spatial Reference ID
      required:
        - x
        - y
        - srid
...
    Error:
      type: object
      description: Fehlermeldung
      properties:
        error:
          type: string
          description: HTTP-Statustext
        message:
          type: string
          description: Detaillierte Fehlermeldung
      required:
        - error
        - message